+++ src/main.rs
@@ -8,6 +8,7 @@ mod config;
 mod i18n;
 mod sftp;
 mod ssh;
+mod ssh_config;
 mod system;
 
 fn main() -> anyhow::Result<()> {